Fix interaction between belongs-to and scoped admin.
Ensure that the currently selected belongs-to association is always in the the list of possible selections because it could be excluded by the scope of normally editable items of that type. For example, a soft-delete feature shouldn't cause you to not see legacy (read deleted) data on a belongs-to--also because you wouldn't want to accidentally unset that relationship just because the scoping couldn't find it when you edited some other unrelated attribute on the object in the admin screen.